+++ tools/Docker/buildDocker.sh
@@ -1,5 +1,7 @@
 #!/bin/sh
+DOCKERHUBID=ankraft
+
 cd ../..
 docker build -t acme-onem2m-cse -f tools/Docker/Dockerfile .
-docker tag acme-onem2m-cse ankraft/acme-onem2m-cse
-docker push ankraft/acme-onem2m-cse
+docker tag acme-onem2m-cse $DOCKERHUBID/acme-onem2m-cse
+docker push $DOCKERHUBID/acme-onem2m-cse