+++ tools/Docker/runDocker.sh
@@ -1,2 +1,4 @@
 #!/bin/sh
-docker run -p 8080:8080 --rm --name acme-onem2m-cse ankraft/acme-onem2m-cse
+DOCKERHUBID=ankraft
+
+docker run -p 8080:8080 --rm --name acme-onem2m-cse $DOCKERHUBID/acme-onem2m-cse