+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [CSE] Discovery supports "attributes + children" return content.
 - [CSE] Changed command line argument --reset-db to --db-reset .
 - [CSE] Added command line argument --db-storage .
+- [CSE] Added support for FlexContainerInstance. 
 
 ## 0.2.1 - 2020-03-06
 - [APPS] Fixed wrong originator handling for already registered AEs.