+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## Unreleased 0.3.0 - 2020-04-20
+## unreleased 0.4.0 - xxxx.xx
+- [WEB] Add displaying the path of the selected resource.
+
+## 0.3.0 - 2020-04-20
 - [CSE] Discovery supports "attributes + children" return content.
 - [CSE] Changed command line argument --reset-db to --db-reset .
 - [CSE] Added command line argument --db-storage .